requirements.txt
fastapi
uvicorn[standard]
onnxruntime
pillow
